app/web/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.23.2
jinja2==3.1.2
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0

# Authentication
jwt==1.3.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client
httpx==0.25.0

# Database
sqlalchemy>=2.0.0
alembic>=1.10.0
psycopg2-binary>=2.9.5
asyncpg>=0.27.0

# Session Management
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
flake8==6.1.0
black==23.10.1 